requirements.txt
asgiref==3.7.2
attrs==23.2.0
backports.zoneinfo==0.2.1
cattrs==23.2.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
distlib==0.3.8
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
filelock==3.13.1
flatbuffers==23.5.26
idna==3.6
mysqlclient==2.2.1
numpy==1.24.4
openmeteo-requests==1.1.0
openmeteo-sdk==1.7.2
panda==0.3.1
pandas==2.0.3
platformdirs==4.1.0
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-cache==1.1.1
retry-requests==2.0.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.4
url-normalize==1.4.3
urllib3==2.1.0
virtualenv==20.25.0
